+++ src/resources/database/migrations/2019_08_10_175251_streamline_provinces_table.php
@@ -4,28 +4,10 @@ declare(strict_types=1);
 
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
-use Illuminate\Support\Facades\DB;
 use Konekt\Address\Models\ProvinceTypeProxy;
 
 class StreamlineProvincesTable extends Migration
 {
-    /**
-     * With Laravel up to v5.8 renaming any column in a table that also has a column
-     * of type enum is not currently supported. This problem only seems to affect
-     * MySQL, this workaround provides a bit dirty but working solution for it
-     *
-     * @see https://laravel.com/docs/5.8/migrations#modifying-columns
-     * @see https://stackoverflow.com/q/33140860/1016746
-     */
-    public function __construct()
-    {
-        $platform = DB::getDoctrineConnection()->getDatabasePlatform();
-
-        if (!$platform->hasDoctrineTypeMappingFor('enum')) {
-            $platform->registerDoctrineTypeMapping('enum', 'string');
-        }
-    }
-
     public function up()
     {
         /** based on ISO 3166-2 https://en.wikipedia.org/wiki/ISO_3166-2 */