+++ Changelog.md
@@ -9,6 +9,7 @@
 - Added PHP 8.3 support
 - Added Laravel 11 support
 - Added Doctrine DBAL v4 support (required for Laravel 11)
+- Removed the Doctrine based migration-workaround for modifying table with enum column that wasn't necessary after Laravel 6
 - Changed minimum Enum version to v4.1
 - BC: Changed all enum interfaces to extend the root `EnumInterface`
 - BC: added the `iso2Code()` and the `getName()` methods to the `Country` interface